react_gen_code/pyproject.toml
[tool.poetry]
name = "react_gen_code"
version = "0.1.0"
description = "Given a yaml file, generates the react code assuming some basic components"
authors = ["Nitin Surya <nitinsurya1991@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^3.0.2"
PyYAML = "^6.0"
munch = "^2.5.0"
watchdog = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
